+++ modules/searchbar/gaz/model.js
@@ -2,6 +2,7 @@ import "../model";
 import store from "../../../src/app-store";
 import {search, setGazetteerUrl, setShowGeographicIdentifier} from "@masterportal/masterportalapi/src/searchAddress";
 import {sort} from "../../../src/utils/sort";
+import uniqueId from "../../../src/utils/uniqueId";
 
 const GazetteerModel = Backbone.Model.extend({
     defaults: {
@@ -106,7 +107,9 @@ const GazetteerModel = Backbone.Model.extend({
             this.pushResult(searchResults[0], "paste");
         }
         else {
-            this.pushAllResults(searchResults);
+            const uniqueResults = this.removeDuplicateAddresses(searchResults);
+
+            this.pushAllResults(uniqueResults);
             Radio.trigger("Searchbar", "createRecommendedList");
         }
     },
@@ -139,10 +142,29 @@ const GazetteerModel = Backbone.Model.extend({
             type: translatedType,
             coordinate: searchResult.geometry.coordinates,
             icon: "bi-signpost-split-fill",
-            id: searchResult.name.replace(/ /g, "") + translatedType,
+            id: uniqueId("gazSuggest"),
             properties: searchResult.properties,
             storedQuery: searchResult.type
         }, evtType);
+    },
+
+    /**
+     * Removes duplicate search results identified by gml:id
+     * @param {Object[]} searchResults The search results.
+
+     * @returns {Object[]} search results without duplicates
+     */
+    removeDuplicateAddresses: function (searchResults) {
+        searchResults.forEach((result, index) => {
+            const isUnique = searchResults.filter(searchResult => result.properties?.$?.["gml:id"] === searchResult.properties?.$?.["gml:id"]).length === 1;
+
+            if (!isUnique) {
+                searchResults.splice(index, 1);
+            }
+            return isUnique;
+        });
+
+        return searchResults;
     },
 
     /**